% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_conditional_logistic_regression}
\alias{is_conditional_logistic_regression}
\title{Interstroke Conditonal Logistic Regression}
\usage{
is_conditional_logistic_regression(interstroke_data, models)
}
\arguments{
\item{interstroke_data}{An Interstroke Dataset}

\item{models}{A list of models}
}
\value{
A dataframe with model output
}
\description{
Interstroke Conditonal Logistic Regression
}
\examples{
is_conditional_logistic_regression(interstroke_data=interstroke_pollution,models=c(case_num~cookplace+strata(caseid),case_num~cookplace+eage+nevfcur+aheiscore+whrs2tert+hhincome+cardiacrfcat+sysbp+education+strata(caseid),case_num~cookplace+eage+nevfcur+aheiscore+whrs2tert+hhincome+cardiacrfcat+sysbp+education+country+incomectry+strata(caseid)))
}
