% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_conditional_logistic_regression_stratified}
\alias{is_conditional_logistic_regression_stratified}
\title{Interstroke conditonal Logistic Regression Stratified}
\usage{
is_conditional_logistic_regression_stratified(
  interstroke_data,
  subgroups,
  model,
  include_overall = TRUE
)
}
\arguments{
\item{interstroke_data}{An Interstroke Dataset}

\item{subgroups}{A list of subgroups for stratified analysis.}

\item{model}{A list of models for table 2}

\item{include_overall}{default is TRUE. Set to FALSE to exclude overall analysis.}
}
\value{
A dataframe with model output
}
\description{
Interstroke conditonal Logistic Regression Stratified
}
\examples{
is_conditional_logistic_regression_stratified(interstroke_data=interstroke_bmi, subgroups=c("esex","smoking"), model=case_num~whr+eage+strata(caseid),include_overall = FALSE)
}
