% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interstroke_forest_plot.R
\name{is_figure_forest_ggplot}
\alias{is_figure_forest_ggplot}
\title{Interstroke forest ggplot}
\usage{
is_figure_forest_ggplot(
  is_forest_data,
  breaks = breaks,
  limits = NULL,
  text_size = NULL,
  title,
  long_title = long_title
)
}
\arguments{
\item{is_forest_data}{Data input from is_forest_figure}

\item{breaks}{manually assign break values for axis.}

\item{limits}{manually assign plot axis limits.}

\item{text_size}{**Optional** Assign value for text size.}

\item{title}{Character string for title of plot.}

\item{long_title}{default TRUE gives a descriptive title. Set to false for just the predictor as title.}
}
\value{
An Interstroke models ggplot figure
}
\description{
Interstroke forest ggplot
}
