% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interstroke_model_figure.R
\name{is_figure_models}
\alias{is_figure_models}
\title{Interstroke Model Figure}
\usage{
is_figure_models(
  interstroke_data,
  models,
  breaks = NULL,
  regression_type = NULL
)
}
\arguments{
\item{interstroke_data}{Project specific Interstroke dataset.}

\item{models}{A list of model formulas to include in model figure (see example)}

\item{breaks}{axis tick mark parameters, defaults to pretty_breaks.}

\item{regression_type}{either "unconditional" or "conditional". Conditional models require +strata(caseid) at end of model formula.}
}
\value{
An Interstroke model figure
}
\description{
Interstroke Model Figure
}
\examples{
is_figure_models(interstroke_data=interstroke_bmi, x_axis_variable=bmi_quintiles, models=c(case_num ~ bmi_quintiles, case_num ~ bmi_quintiles+eage+esex))
}
