% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interstroke_model_figure.R
\name{is_figure_models_ggplot}
\alias{is_figure_models_ggplot}
\title{Interstroke Model ggplot}
\usage{
is_figure_models_ggplot(
  interstroke_modelling_data,
  breaks = breaks,
  x_axis_variable = x_axis_variable,
  interstroke_data = interstroke_data
)
}
\arguments{
\item{interstroke_modelling_data}{data imported from is figure models function.}

\item{breaks}{breaks imported from is figure models function.}
}
\value{
An Interstroke models ggplot figure
}
\description{
Interstroke Model ggplot
}
