% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interstroke_mrs_figure.R
\name{is_figure_mrs}
\alias{is_figure_mrs}
\title{Interstroke MRS Figure}
\usage{
is_figure_mrs(
  interstroke_data,
  x_axis_variable,
  dropped_prop_levels = NULL,
  na_rm = FALSE,
  text_size = NULL
)
}
\arguments{
\item{interstroke_data}{Project specific Interstroke dataset.}

\item{x_axis_variable}{A categorical variable that will be placed on the x-axis of the proportion figure.}

\item{dropped_prop_levels}{**optional** Named proportion levels to be excluded. Default includes all predictor levels. Format = c("x", "y", "z")}

\item{na_rm}{**optional** A logical scalar. Should NA values be removed?. Default is FALSE.}

\item{text_size}{**Optional** Assign value for text size.}
}
\value{
An Interstroke MRS ggplot figure
}
\description{
Interstroke MRS Figure
}
\examples{
is_figure_mrs(interstroke_sleep, csdiffsl)
is_figure_mrs(interstroke_sleep \%>\% drop_na(csdiffsl) \%>\% filter(case== "Case"),csdiffsl)
is_figure_mrs(interstroke_proxy, regionnn7, highwhr)
is_figure_mrs(interstroke_proxy, regionnn7, highwhr, dropped_prop_levels = c("Low", "Moderate"), na_rm = TRUE)
}
