% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interstroke_proportion_figure.R
\name{is_figure_proportion}
\alias{is_figure_proportion}
\title{Interstroke Proportion Figure}
\usage{
is_figure_proportion(
  interstroke_data,
  x_axis_variable,
  proportion_variable,
  dropped_prop_levels = NULL,
  na_rm = FALSE,
  text_size = NULL
)
}
\arguments{
\item{interstroke_data}{Project specific Interstroke dataset.}

\item{x_axis_variable}{A categorical variable that will be placed on the x-axis of the proportion figure.}

\item{proportion_variable}{The predictor variable.}

\item{dropped_prop_levels}{**optional** Named proportion levels to be excluded. Default includes all predictor levels. Format = c("x", "y", "z")}

\item{na_rm}{**optional** A logical scalar. Should NA values be removed?. Default is FALSE.}

\item{text_size}{**Optional** Assign value for text size.}
}
\value{
An Interstroke proportion ggplot figure
}
\description{
Interstroke Proportion Figure
}
\examples{
is_figure_proportion(interstroke_bmi, regionnn7, bmi_3_level)
is_figure_proportion(interstroke_proxy, regionnn7, highwhr)
is_figure_proportion(interstroke_proxy, regionnn7, highwhr, dropped_prop_levels = c("Low", "Moderate"), na_rm = TRUE)
}
