% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interstroke_spline_figure.R
\name{is_figure_spline}
\alias{is_figure_spline}
\title{Interstroke Spline Figure}
\usage{
is_figure_spline(interstroke_data, formula, predictor, env = caller_env())
}
\arguments{
\item{interstroke_data}{Project specific Interstroke dataset.}

\item{formula}{Model formulas to include in figure (see example)}

\item{predictor}{Predictor variable.}
}
\value{
An Interstroke spline figure
}
\description{
Interstroke Spline Figure
}
\examples{
is_figure_spline(interstroke_data=interstroke_pollution, formula=case_num ~ rcs(facthmkm,3),predictor=facthmkm)
}
