% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interstroke_forest_plot.R
\name{is_figure_stratified_subgroups}
\alias{is_figure_stratified_subgroups}
\title{Interstroke stratified analysis for predictor}
\usage{
is_figure_stratified_subgroups(
  interstroke_data,
  subgroups,
  model,
  limits = NULL,
  p_sig = NULL,
  breaks = NULL,
  text_size = NULL,
  drop_columns = NULL,
  include_overall = TRUE,
  long_title = TRUE,
  regression_type = NULL
)
}
\arguments{
\item{interstroke_data}{Project specific Interstroke dataset.}

\item{subgroups}{A list of subgroups for stratified analysis.}

\item{model}{Model formula for analysis. Predictor should be first independant variable.}

\item{limits}{**Optional** Limit the x axis in the format c(xmin, xmax).}

\item{p_sig}{**Optional** Set the level of significance to hide non significant values. Default displays all p values <1.}

\item{breaks}{**Optional** Set the sequence for breaks in the format seq(xmin, xmax, interval)}

\item{text_size}{**Optional** Assign value for text size.}

\item{drop_columns}{**Optional** Drop population and subgroup columns suitable for joint plots.}

\item{long_title}{**Optional** Set to false to just display predictor as title}

\item{regression_type}{either "unconditional" or "conditional". Conditional models require +strata(caseid) at end of model formula.}
}
\value{
A dataframe suitable for Interstroke forest ggplot function.
}
\description{
Interstroke stratified analysis for predictor
}
\examples{
is_figure_stratified_subgroups(interstroke_data=interstroke_bmi, subgroups=c("esex","smoking","alcohfreqcat","antihypertpre","cardiacrfcat"), model=case_num~whr+esex+smokingregression_type = "unconditional")
}
