% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interstroke_subtypes.R
\name{is_figure_subtypes}
\alias{is_figure_subtypes}
\title{Interstroke subtypes Figure}
\usage{
is_figure_subtypes(
  interstroke_data,
  model,
  breaks = NULL,
  regression_type = NULL
)
}
\arguments{
\item{interstroke_data}{Project specific Interstroke dataset.}

\item{model}{A list of model formulas to include in model figure (see example)}

\item{breaks}{Manually assign breaks for axis ticks.}

\item{regression_type}{Either "unconditional" or "conditional". Conditional models require +strata(caseid) at end of model formula.}
}
\value{
An Interstroke model figure
}
\description{
Interstroke subtypes Figure
}
\examples{
is_figure_subtypes(interstroke_data=interstroke_bmi, model=case_num ~ bmi_quintiles)
}
