% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_logistic_regression}
\alias{is_logistic_regression}
\title{Interstroke Logistic Regression}
\usage{
is_logistic_regression(interstroke_data, regression_type, models)
}
\arguments{
\item{interstroke_data}{An Interstroke Dataset}

\item{regression_type}{either "unconditional" or "conditional". Conditional models require +strata(caseid) at end of model formula.}

\item{models}{A list of models. Format = c(case_num~predictor+optional_covariate)}
}
\value{
A dataframe with model output
}
\description{
Interstroke Logistic Regression
}
\examples{
is_logistic_regression(interstroke_data=interstroke_bmi, regression_type = "unconditional", models=c(case_num~whr_z,case_num~whr_z+eage+nevfcur+aheiscore+whrs2tert+hhincome+cardiacrfcat+sysbp,case_num~whr_z+eage+nevfcur+aheiscore+whrs2tert+hhincome+cardiacrfcat+sysbp+incomectry))
is_logistic_regression(interstroke_data=interstroke_pollution, regression_type = "conditional", models=c(case_num~cookplace+strata(caseid),case_num~cookplace+eage+nevfcur+aheiscore+whrs2tert+hhincome+cardiacrfcat+sysbp+education+strata(caseid),case_num~cookplace+eage+nevfcur+aheiscore+whrs2tert+hhincome+cardiacrfcat+sysbp+education+country+incomectry+strata(caseid)))
}
