% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_logistic_regression_stratified}
\alias{is_logistic_regression_stratified}
\title{Interstroke Logistic Regression Stratified}
\usage{
is_logistic_regression_stratified(
  interstroke_data,
  regression_type,
  subgroups,
  model,
  include_overall = TRUE
)
}
\arguments{
\item{interstroke_data}{An Interstroke Dataset}

\item{regression_type}{"unconditional" or "conditional". Include +strata(caseid) in model formula for conditional analysis.}

\item{subgroups}{A list of subgroups for stratified analysis.}

\item{model}{A list of models for table 2}

\item{include_overall}{default is TRUE. Set to FALSE to exclude overall analysis.}
}
\value{
A dataframe with model output
}
\description{
Interstroke Logistic Regression Stratified
}
\examples{
is_logistic_regression_stratified(interstroke_data=interstroke_bmi, regression_type = "unconditional", subgroups=c("esex","smoking"), model=case_num~whr_z+eage)
is_logistic_regression_stratified(interstroke_data=interstroke_bmi, regression_type = "conditional", subgroups=c("esex","smoking"), model=case_num~bmi_z+eage+strata(caseid),include_overall = FALSE)
}
