% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_model_label}
\alias{is_model_label}
\title{Create character string of model}
\usage{
is_model_label(interstroke_data, model_formula)
}
\arguments{
\item{interstroke_data}{interstroke dataset}

\item{model_formula}{A formula to extract the label for}
}
\value{
An Interstroke Dataset
}
\description{
Create character string of model
}
\examples{
is_model_label(interstroke_data = interstroke_bmi, model_formula = case_num ~ cookplace+eage+esex)
is_model_label(interstroke_data = interstroke_bmi, model_formula = case_num ~ cookplace+eage+esex+strata(caseid))
}
