% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interstroke_par.R
\name{is_par}
\alias{is_par}
\title{Interstroke PAR function for table input}
\usage{
is_par(interstroke_data)
}
\arguments{
\item{interstroke_data}{An Interstroke Dataset}
}
\value{
A dataframe with model output
}
\description{
Interstroke PAR function for table input
}
\examples{
is_par()
}
