% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_spline}
\alias{is_spline}
\title{Interstroke Restricted Cubic Splines}
\usage{
is_spline(interstroke_data, model, env = rlang::caller_env())
}
\arguments{
\item{interstroke_data}{An Interstroke Dataset}

\item{model}{Formula for model e.g. case_num ~ rcs(trafhmkm,3)}
}
\value{
A restricted cubic spline plot
}
\description{
Interstroke Restricted Cubic Splines
}
\examples{
is_spline(interstroke_bmi, case_num ~ whr_z)
}
