% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_subset}
\alias{is_subset}
\title{Subset Interstroke and keep case/control match}
\usage{
is_subset(interstroke_data, subset_rule)
}
\arguments{
\item{interstroke_data}{An Interstroke Dataset}

\item{subset_rule}{A subsetting rule e.g. "age <= 45"}
}
\value{
An Interstroke Dataset
}
\description{
Subset Interstroke and keep case/control match
}
\examples{
is_subset(interstroke_data = interstroke_stress, subset_rule = eage <= 45)
}
