% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_table_one}
\alias{is_table_one}
\title{Interstroke Table one}
\usage{
is_table_one(
  interstroke_data,
  model,
  table_title = NULL,
  env = rlang::caller_env()
)
}
\arguments{
\item{interstroke_data}{An Interstroke Dataset}

\item{model}{Standard R format formula e.g. predictor ~ covariate1 + covariate2}

\item{table_title}{Table title}
}
\value{
A table one flextable
}
\description{
Interstroke Table one
}
\examples{
is_table_one(interstroke_data=interstroke_pollution, model = polluteh~eage+bmi+sysbp, table_title="Table 1 - Baseline Characteristics by pollution level")
}
