% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_table_two}
\alias{is_table_two}
\title{Interstroke Table two}
\usage{
is_table_two(
  interstroke_data,
  models,
  regression_type = NULL,
  table_title = NULL
)
}
\arguments{
\item{interstroke_data}{An Interstroke Dataset}

\item{models}{A list of models for table 2}

\item{regression_type}{Type of regression, either conditional or unconditional}

\item{table_title}{Table title}
}
\value{
A table two flextable
}
\description{
Interstroke Table two
}
\examples{
is_table_two(interstroke_data=interstroke_pollution, regression_type="unconditional", models=c(case_num~cookplace,case_num~cookplace+eage+nevfcur+aheiscore+whrs2tert+hhincome+cardiacrfcat+sysbp+education,case_num~cookplace+eage+nevfcur+aheiscore+whrs2tert+hhincome+cardiacrfcat+sysbp+education+country+incomectry))
is_table_two(interstroke_data=interstroke_pollution, regression_type="conditional", models=c(case_num~cookplace+strata(caseid),case_num~cookplace+eage+nevfcur+aheiscore+whrs2tert+hhincome+cardiacrfcat+sysbp+education+strata(caseid),case_num~cookplace+eage+nevfcur+aheiscore+whrs2tert+hhincome+cardiacrfcat+sysbp+education+country+incomectry+strata(caseid)),table_title="Custom defined title using table_title")
}
