% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_tb_prevalence}
\alias{is_tb_prevalence}
\title{is_tb_prevalence A function for adding prevalence to Interstroke tables}
\usage{
is_tb_prevalence(interstroke_data, predictor)
}
\arguments{
\item{interstroke_data}{An Interstroke Dataset}

\item{predictor}{The predictor variable}
}
\value{
A tibble containing a count of Controls, Ischemic and ICH (strktype) at each level of a predictor
}
\description{
is_tb_prevalence A function for adding prevalence to Interstroke tables
}
\examples{
is_tb_prevalence(interstroke_pollution,polluteh)
}
