% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_tb_univariate}
\alias{is_tb_univariate}
\title{is_tb_univariate A function for adding univariate analysis to Interstroke tables}
\usage{
is_tb_univariate(
  interstroke_data,
  predictor,
  regression_type = NULL,
  env = rlang::caller_env()
)
}
\arguments{
\item{interstroke_data}{An Interstroke Dataset}

\item{predictor}{The predictor variable}

\item{regression_type}{"unconditional" or "conditional"}
}
\value{
A tibble containing a count of Controls, Ischemic and ICH (strktype) at each level of a predictor
}
\description{
is_tb_univariate A function for adding univariate analysis to Interstroke tables
}
\examples{
is_tb_univariate(interstroke_pollution,polluteh, regression_type = "unconditional")
is_tb_univariate(interstroke_pollution,polluteh, regression_type = "conditional")
}
