% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_tb_univariate_conditonal}
\alias{is_tb_univariate_conditonal}
\title{is_tb_univariate_conditonal A function for adding univariate conditonal analysis to Interstroke tables}
\usage{
is_tb_univariate_conditonal(
  interstroke_data,
  predictor,
  env = rlang::caller_env()
)
}
\arguments{
\item{interstroke_data}{An Interstroke Dataset}

\item{predictor}{The predictor variable}
}
\value{
A tibble containing a count of Controls, Ischemic and ICH (strktype) at each level of a predictor
}
\description{
is_tb_univariate_conditonal A function for adding univariate conditonal analysis to Interstroke tables
}
\examples{
is_tb_univariate_conditonal(interstroke_pollution,polluteh)
}
