% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_tb_univariate_unconditonal}
\alias{is_tb_univariate_unconditonal}
\title{is_tb_univariate_unconditonal A function for adding univariate unconditonal analysis to Interstroke tables}
\usage{
is_tb_univariate_unconditonal(
  interstroke_data,
  predictor,
  predictor_name,
  env = rlang::caller_env()
)
}
\arguments{
\item{interstroke_data}{An Interstroke Dataset}

\item{predictor}{The predictor variable}
}
\value{
A tibble containing a count of Controls, Ischemic and ICH (strktype) at each level of a predictor
}
\description{
is_tb_univariate_unconditonal A function for adding univariate unconditonal analysis to Interstroke tables
}
\examples{
is_tb_univariate_unconditonal(interstroke_pollution,polluteh)
}
